+++ flink-connectors/flink-connector-aws-kinesis-data-streams/src/test/java/org/apache/flink/streaming/connectors/kinesis/testutils/KinesaliteContainer.java
@@ -25,11 +25,22 @@ import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.services.kinesis.AmazonKinesis;
 import com.amazonaws.services.kinesis.AmazonKinesisClientBuilder;
 import com.amazonaws.services.kinesis.model.ListStreamsResult;
+import org.rnorth.ducttape.ratelimits.RateLimiter;
+import org.rnorth.ducttape.ratelimits.RateLimiterBuilder;
 import org.rnorth.ducttape.unreliables.Unreliables;
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.wait.strategy.AbstractWaitStrategy;
 import org.testcontainers.utility.DockerImageName;
+import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
+import software.amazon.awssdk.http.SdkHttpConfigurationOption;
+import software.amazon.awssdk.http.async.SdkAsyncHttpClient;
+import software.amazon.awssdk.http.nio.netty.NettyNioAsyncHttpClient;
+import software.amazon.awssdk.regions.Region;
+import software.amazon.awssdk.services.kinesis.KinesisAsyncClient;
+import software.amazon.awssdk.utils.AttributeMap;
 
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.util.Properties;
 import java.util.concurrent.TimeUnit;
 
@@ -45,6 +56,7 @@ import static com.amazonaws.SDKGlobalConfiguration.SECRET_KEY_ENV_VAR;
 public class KinesaliteContainer extends GenericContainer<KinesaliteContainer> {
     private static final String ACCESS_KEY = "access key";
     private static final String SECRET_KEY = "secret key";
+    private static final Region REGION = Region.US_EAST_1;
 
     public KinesaliteContainer(DockerImageName imageName) {
         super(imageName);
@@ -53,15 +65,7 @@ public class KinesaliteContainer extends GenericContainer<KinesaliteContainer> {
         withEnv(SECRET_KEY_ENV_VAR, ACCESS_KEY);
         withExposedPorts(4567);
         waitingFor(new ListStreamsWaitStrategy());
-        withCreateContainerCmdModifier(
-                cmd ->
-                        cmd.withEntrypoint(
-                                "/tini",
-                                "--",
-                                "/usr/src/app/node_modules/kinesalite/cli.js",
-                                "--path",
-                                "/var/lib/kinesalite",
-                                "--ssl"));
+        tryStartContainer();
     }
 
     /** Returns the endpoint url to access the container from outside the docker network. */
@@ -80,6 +84,10 @@ public class KinesaliteContainer extends GenericContainer<KinesaliteContainer> {
 
     public String getSecretKey() {
         return SECRET_KEY;
+    }
+
+    public Region getRegion() {
+        return REGION;
     }
 
     /** Returns the properties to access the container from outside the docker network. */
@@ -112,9 +120,31 @@ public class KinesaliteContainer extends GenericContainer<KinesaliteContainer> {
                 .build();
     }
 
+    public KinesisAsyncClient getV2Client() throws URISyntaxException {
+        return KinesisAsyncClient.builder()
+                .endpointOverride(new URI(getHostEndpointUrl()))
+                .region(REGION)
+                .credentialsProvider(
+                        () -> AwsBasicCredentials.create(getAccessKey(), getSecretKey()))
+                .httpClient(buildSdkAsyncHttpClient())
+                .build();
+    }
+
+    private void tryStartContainer() {
+        withCreateContainerCmdModifier(
+                cmd ->
+                        cmd.withEntrypoint(
+                                "/tini",
+                                "--",
+                                "/usr/src/app/node_modules/kinesalite/cli.js",
+                                "--path",
+                                "/var/lib/kinesalite",
+                                "--ssl"));
+    }
+
     private Properties getProperties(String endpointUrl) {
         Properties config = new Properties();
-        config.setProperty(AWSConfigConstants.AWS_REGION, "us-east-1");
+        config.setProperty(AWSConfigConstants.AWS_REGION, REGION.toString());
         config.setProperty(AWSConfigConstants.AWS_ENDPOINT, endpointUrl);
         config.setProperty(AWSConfigConstants.AWS_ACCESS_KEY_ID, getAccessKey());
         config.setProperty(AWSConfigConstants.AWS_SECRET_ACCESS_KEY, getSecretKey());
@@ -122,16 +152,33 @@ public class KinesaliteContainer extends GenericContainer<KinesaliteContainer> {
     }
 
     private class ListStreamsWaitStrategy extends AbstractWaitStrategy {
+        private static final int TIMEOUT_PER_RETRY = 15;
+
+        private final RateLimiter rateLimiter =
+                RateLimiterBuilder.newBuilder()
+                        .withRate(TIMEOUT_PER_RETRY, TimeUnit.SECONDS)
+                        .withConstantThroughput()
+                        .build();
+
         @Override
         protected void waitUntilReady() {
             Unreliables.retryUntilSuccess(
-                    (int) this.startupTimeout.getSeconds(),
+                    (int) this.startupTimeout.getSeconds() * TIMEOUT_PER_RETRY,
                     TimeUnit.SECONDS,
-                    () -> this.getRateLimiter().getWhenReady(() -> tryList()));
+                    () -> rateLimiter.getWhenReady(() -> tryList()));
         }
 
         private ListStreamsResult tryList() {
+            tryStartContainer();
             return getContainerClient().listStreams();
         }
+    }
+
+    private SdkAsyncHttpClient buildSdkAsyncHttpClient() {
+        return NettyNioAsyncHttpClient.builder()
+                .buildWithDefaults(
+                        AttributeMap.builder()
+                                .put(SdkHttpConfigurationOption.TRUST_ALL_CERTIFICATES, true)
+                                .build());
     }
 }
